Undefined symbol fixed, segfault now

+++ transformations/interpolation/InterpExpo.cc
@@ -15,6 +15,8 @@ using std::prev;
 using std::advance;
 using TMath::Exp;
 
+
+
 //InterpExpo::InterpExpo(const std::string& underflow_strategy, const std::string& overflow_strategy) : InSegment() {
 
 InterpExpo::InterpExpo() : InSegment() {
@@ -168,7 +170,7 @@ void InterpExpo::do_interpolate_ongpu(FunctionArgs& fargs) {
 	fargs.gpu->arg[0].size(), fargs.gpu->arg[1].size()) // TODO rewrite?
 */
         auto& gpuargs=fargs.gpu;
-	interpExpo_v1(gpuargs->args, gpuargs->rets, gpuargs->argshapes[0][0], gpuargs->argshapes[1][0] );
+	interpExpo_v1(gpuargs->args, gpuargs->rets, (int)gpuargs->argshapes[0][0], (int)gpuargs->argshapes[1][0] );
 /*    interpExpo_v1(gpuargs->args[0], gpuargs->rets[0], gpuargs->args[1], 
 	gpuargs->args[4], gpuargs->args[2], gpuargs->args[3], 
 	gpuargs->argshapes[0][0], gpuargs->argshapes[1][0]); // TODO rewrite?
